custom dialog for setup second device (add steps)
to address #3178
