+++ src/renderer/components/dialogs/SetupMultiDevice/index.ts
@@ -1,2 +1,3 @@
 export { SendBackupDialog } from './SendBackupDialog'
 export { ReceiveBackupDialog } from './ReceiveBackupDialog'
+export { ReceiveBackupProgressDialog } from './ReceiveBackupProgressDialog'
